Move links outside
